+++ src/main/java/com/scubakay/autorelog/commands/AutoRelogCommand.java
@@ -6,6 +6,9 @@ import com.mojang.brigadier.context.CommandContext;
 import com.mojang.brigadier.tree.ArgumentCommandNode;
 import com.mojang.brigadier.tree.LiteralCommandNode;
 import com.scubakay.autorelog.AutoRelogClient;
+import com.scubakay.autorelog.commands.suggestions.DelaySuggestionProvider;
+import com.scubakay.autorelog.commands.suggestions.IntervalSuggestionProvider;
+import com.scubakay.autorelog.commands.suggestions.MaxAttemptsSuggestionProvider;
 import com.scubakay.autorelog.util.Reconnect;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
 import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
@@ -21,22 +24,36 @@ public class AutoRelogCommand {
 
         LiteralCommandNode<FabricClientCommandSource> delayNode = ClientCommandManager
                 .literal("delay")
-                .executes(AutoRelogCommand::autorelog)
+                .build();
+
+        LiteralCommandNode<FabricClientCommandSource> configNode = ClientCommandManager
+                .literal("config")
                 .build();
 
         ArgumentCommandNode<FabricClientCommandSource, Integer> delayArgumentNode = ClientCommandManager
                 .argument("delay", IntegerArgumentType.integer())
+                .suggests(new DelaySuggestionProvider())
                 .executes(ctx -> delay(ctx, IntegerArgumentType.getInteger(ctx, "delay")))
                 .build();
 
         LiteralCommandNode<FabricClientCommandSource> intervalNode = ClientCommandManager
                 .literal("interval")
-                .executes(AutoRelogCommand::autorelog)
                 .build();
 
         ArgumentCommandNode<FabricClientCommandSource, Integer> intervalArgumentNode = ClientCommandManager
                 .argument("interval", IntegerArgumentType.integer())
+                .suggests(new IntervalSuggestionProvider())
                 .executes(ctx -> interval(ctx, IntegerArgumentType.getInteger(ctx, "interval")))
+                .build();
+
+        LiteralCommandNode<FabricClientCommandSource> maxAttemptsNode = ClientCommandManager
+                .literal("maxAttempts")
+                .build();
+
+        ArgumentCommandNode<FabricClientCommandSource, Integer> maxAttemptsArgumentNode = ClientCommandManager
+                .argument("maxAttempts", IntegerArgumentType.integer())
+                .suggests(new MaxAttemptsSuggestionProvider())
+                .executes(ctx -> maxAttempts(ctx, IntegerArgumentType.getInteger(ctx, "maxAttempts")))
                 .build();
 
         LiteralCommandNode<FabricClientCommandSource> cancelNode = ClientCommandManager
@@ -46,13 +63,20 @@ public class AutoRelogCommand {
 
         dispatcher.getRoot().addChild(autoRelogNode);
 
+        // Add config node
+        autoRelogNode.addChild(configNode);
+
         // Add delay config node
-        autoRelogNode.addChild(delayNode);
+        configNode.addChild(delayNode);
         delayNode.addChild(delayArgumentNode);
 
         // Add delay config node
-        autoRelogNode.addChild(intervalNode);
+        configNode.addChild(intervalNode);
         intervalNode.addChild(intervalArgumentNode);
+
+        // Add max attempts config node
+        configNode.addChild(maxAttemptsNode);
+        maxAttemptsNode.addChild(maxAttemptsArgumentNode);
 
         autoRelogNode.addChild(cancelNode);
     }
@@ -60,7 +84,7 @@ public class AutoRelogCommand {
     private static int autorelog(CommandContext<FabricClientCommandSource> context) {
         if (context.getSource().getClient().isInSingleplayer()) {
             context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_error_singleplayer"), false);
-            return 1;
+            return -1;
         }
         Reconnect.getInstance().activate();
         context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_activated"), false);
@@ -70,7 +94,7 @@ public class AutoRelogCommand {
     private static int cancel(CommandContext<FabricClientCommandSource> context) {
         if (context.getSource().getClient().isInSingleplayer()) {
             context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_error_singleplayer"), false);
-            return 1;
+            return -1;
         }
         Reconnect.getInstance().deactivate();
         context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_deactivated"), false);
@@ -78,14 +102,32 @@ public class AutoRelogCommand {
     }
 
     private static int delay(CommandContext<FabricClientCommandSource> context, int delay) {
+        if (delay < 1) {
+            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_delay"), false);
+            return -1;
+        }
         AutoRelogClient.CONFIG.setDelay(delay);
         context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_delay_changed", delay), false);
         return 1;
     }
 
     private static int interval(CommandContext<FabricClientCommandSource> context, int interval) {
+        if (interval < 1) {
+            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_interval"), false);
+            return -1;
+        }
         AutoRelogClient.CONFIG.setInterval(interval);
         context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_interval_changed", interval), false);
+        return 1;
+    }
+
+    private static int maxAttempts(CommandContext<FabricClientCommandSource> context, int maxAttempts) {
+        if (maxAttempts < 0) {
+            context.getSource().getPlayer().sendMessage(Text.translatable("commands.config_error_max_attempts"), false);
+            return -1;
+        }
+        AutoRelogClient.CONFIG.setMaxAttempts(maxAttempts);
+        context.getSource().getPlayer().sendMessage(Text.translatable("commands.autorelog_max_attempts_changed", maxAttempts), false);
         return 1;
     }
 }